refactor: according to PR notions
