+++ .env.example
@@ -1,0 +1,2 @@
+VITE_SUPABASE_URL='https://your_supabase_url.supabase.co'
+VITE_SUPABASE_ANON_KEY='your_supabase_anon_key'
