+++ .gitignore
@@ -24,3 +24,6 @@ coverage
 *.njsproj
 *.sln
 *.sw?
+
+
+.env
