--- src/pages/login/LogInPage.tsx
@@ -1,5 +1,0 @@
-function LoginPage() {
-  return <h1>Login Page</h1>;
-}
-
-export default LoginPage;
